feat (client-api): get correct url for zeppelin and minio when running locally
