+++ code/workspaces/client-api/src/dataaccess/stackUrlService.js
@@ -1,9 +1,8 @@
-import { findLast } from 'lodash';
-import axios from 'axios';
 import logger from 'winston';
 import { stackTypes } from 'common';
 import rstudioTokenService from './login/rstudioTokenService';
 import vault from './vault/vault';
+import zeppelinLogin from './login/zeppelinLogin';
 
 // NOTE: All other stack details should come from 'common/src/config/images'
 const { JUPYTER, JUPYTERLAB, ZEPPELIN, RSTUDIO, NBVIEWER } = stackTypes;
@@ -64,31 +63,3 @@ function requestRStudioToken(projectKey, stack) {
     });
 }
 
-const zeppelinLogin = notebook => (credentials) => {
-  if (!credentials.username || !credentials.password) {
-    return Promise.reject(new Error('username or password not defined'));
-  }
-
-  return axios.post(getZeppelinLoginUrl(notebook), createLoginPayload(credentials))
-    .then(processLoginResponse);
-};
-
-function getZeppelinLoginUrl(notebook) {
-  const zeppelinUrl = `${notebook.internalEndpoint}/api/login`;
-  logger.info(`Request log in cookie from Zeppelin at: ${zeppelinUrl}`);
-  return zeppelinUrl;
-}
-
-function createLoginPayload(credentials) {
-  return `userName=${credentials.username}&password=${credentials.password}`;
-}
-
-function processLoginResponse(loginResponse) {
-  if (loginResponse.data.error) {
-    return Promise.reject(new Error(`Zeppelin login failed ${loginResponse.data.error.message}`));
-  }
-
-  const { headers } = loginResponse;
-  const sessionCookie = findLast(headers['set-cookie'], header => header.indexOf('JSESSIONID') > -1);
-  return sessionCookie.split(';')[0].split('=')[1];
-}
